refractor: moved library to src
